refactor(services/oas): put v2 builders into v2 folder; use dropdown for spec definitions
